+++ include/matrix.h
@@ -65,10 +65,10 @@ typedef struct {
   // ADC value when the key is fully pressed
   uint16_t adc_bottom_out_value;
 
-  // Key travel distance (0-255)
-  uint8_t distance;
-  // Last extremum point of the key travel distance (0-255)
-  uint8_t extremum;
+  // Key travel distance (0-10000)
+  uint16_t distance;
+  // Last extremum point of the key travel distance (0-10000)
+  uint16_t extremum;
   // Current key travel direction
   uint8_t key_dir;
   // Whether the key is pressed
